back up

+++ best_model_test.py
@@ -6,7 +6,7 @@ current_path = os.path.dirname(__file__)
 
 test_data = data_util.read_variable('testing_data.pickle')
 t2s_test_pad = test_data[0]
-Y_test = test_data[1]
+Y_test = test_data[1].astype(int)
 model = load_model('sequential_1.h5')
 model.compile(loss="binary_crossentropy", optimizer="adam", metrics=["accuracy"])
 y_pred_p = model.predict(t2s_test_pad)
